Revert "merge with gpl 386.52294, add support for rt-ax3000n"
This reverts commit 6f565ac05e145ee988be3a6323f47c6c6cf487de.

+++ release/src/router/strongswan/src/libstrongswan/collections/array.c
@@ -1,8 +1,9 @@
 /*
  * Copyright (C) 2014 Tobias Brunner
- * Copyright (C) 2013 Martin Willi
+ * HSR Hochschule fuer Technik Rapperswil
  *
- * Copyright (C) secunet Security Networks AG
+ * Copyright (C) 2013 Martin Willi
+ * Copyright (C) 2013 revosec AG
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -67,7 +68,7 @@ static size_t get_size(array_t *array, uint32_t num)
 {
 	if (array->esize)
 	{
-		return (size_t)array->esize * num;
+		return array->esize * num;
 	}
 	return sizeof(void*) * num;
 }
